requirements.txt
# UI
streamlit==1.48.0

# .docx read/write + XML (comments)
python-docx==1.2.0
lxml==5.2.2

# RAG framework + vector DB + embeddings
langchain==0.3.27
chromadb==1.0.16
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
tokenizers>=0.19.1

# Parsing PDFs / docs (for ADGM reference materials)
pypdf==4.2.0
unstructured==0.15.0

# Runtime utils
pydantic==2.7.4
python-dotenv==1.0.1
requests>=2.32.3
PyYAML>=6.0.2
tiktoken==0.7.0
